JWG:  Organizing into packages for multiple years
