+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/centerstage/AutonBlueLeft.java
@@ -1,16 +1,17 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.centerstage;
 
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.util.ElapsedTime;
 
-
 @Autonomous
-public class AutonRedRightTest extends LinearOpMode {
-    SpikeCam.location mySpikeLocation;
+public class AutonBlueLeft extends LinearOpMode {
+// This is a SHORT side Auton
+SpikeCam.location mySpikeLocation;
+    CyDogsSparky mySparky;
 
-    // This is SHORT side Auton test
+    // This is a SHORT side Auton
     @Override
     public void runOpMode() {
 
@@ -22,10 +23,10 @@ public class AutonRedRightTest extends LinearOpMode {
         CyDogsChassis.Direction parkingSpot = CyDogsChassis.Direction.LEFT;
 
         // Create the instance of sparky, initialize the SpikeCam, devices, and positions
-        CyDogsSparky mySparky = new CyDogsSparky(this, CyDogsChassis.Alliance.RED, 330);
+        mySparky = new CyDogsSparky(this, CyDogsChassis.Alliance.BLUE,350);
         mySparky.initializeSpikeCam();
         mySparky.initializeDevices();
- //       mySparky.initializePositions();
+
         mySparky.initializeAprilTags();
 
         // Ask the initialization questions
@@ -42,63 +43,58 @@ public class AutonRedRightTest extends LinearOpMode {
 
             // Get to standard position before placing purple pixel
             mySparky.MoveStraight(-300, .5, mySparky.StandardAutonWaitTime);
-            mySparky.StrafeRight(90,0.5, mySparky.StandardAutonWaitTime);
+            mySparky.StrafeLeft(75,0.5, mySparky.StandardAutonWaitTime);
             mySparky.MoveStraight(-445, .5, mySparky.StandardAutonWaitTime);
 
             // Place purple pixel and back away from it
             mySparky.AutonPlacePurplePixel(mySpikeLocation);
-            if(mySpikeLocation== SpikeCam.location.MIDDLE) {
-                mySparky.raiseArmToScore(CyDogsSparky.ArmRaiseBeforeElbowMovement);
+            if(mySpikeLocation== SpikeCam.location.LEFT) {
                 mySparky.MoveStraight(65, .5, mySparky.StandardAutonWaitTime);
+                mySparky.raiseArmToScore(CyDogsSparky.ArmRaiseBeforeElbowMovement);
+                sleep(400);
             }
             else {
-                mySparky.MoveStraight(20, .5, mySparky.StandardAutonWaitTime);
+                mySparky.MoveStraight(30, .5, mySparky.StandardAutonWaitTime);
             }
+
 
             // First, let's get ourselves straight facing scoring area
             //   Then, adjust position.  Remember dropping purple pixel moved us back from spike 20mm
             if (mySpikeLocation == SpikeCam.location.LEFT) {
-                //Already facing the correct way
-                //We're 'BackUpDistanceFromSpike' closer to scoreboard
-                mySparky.RotateRight(3,.5,mySparky.StandardAutonWaitTime);
-                mySparky.StrafeRight(40,.5,mySparky.StandardAutonWaitTime);
-                mySparky.raiseArmToScore(CyDogsSparky.ArmRaiseBeforeElbowMovement);
-                // has long wait time to handle arm movement before it
-                mySparky.MoveStraight(675, .5, 2000);
-      // I took 200 off the above to be far enough away to read april tags
+
+                mySparky.StrafeRight(CyDogsChassis.OneTileMM, .5, mySparky.StandardAutonWaitTime);
+                mySparky.MoveStraight(-CyDogsChassis.OneTileMM-160, .5, mySparky.StandardAutonWaitTime);
+                mySparky.StrafeLeft(CyDogsChassis.OneTileMM-40, .5, mySparky.StandardAutonWaitTime);
+
+                mySparky.RotateLeft(188, .5, 2000);
             } else if (mySpikeLocation == SpikeCam.location.MIDDLE) {
-                mySparky.RotateLeft(92, .5, mySparky.StandardAutonWaitTime);
+                mySparky.RotateRight(92, .5, mySparky.StandardAutonWaitTime);
                 // We're 50mm further away from start position
-                mySparky.StrafeRight(-50,.5,mySparky.StandardAutonWaitTime);
-               // mySparky.raiseArmToScore(CyDogsSparky.ArmRaiseBeforeElbowMovement);
-                // has long wait time to handle arm movement before it
+                mySparky.StrafeLeft(-50,.5,mySparky.StandardAutonWaitTime);
+                mySparky.raiseArmToScore(CyDogsSparky.ArmRaiseBeforeElbowMovement);
                 mySparky.MoveStraight(500, .5, 2000);
             } else {  //RIGHT
-                mySparky.StrafeLeft(CyDogsChassis.OneTileMM, .5, mySparky.StandardAutonWaitTime);
-                mySparky.MoveStraight(-CyDogsChassis.OneTileMM-160, .5, mySparky.StandardAutonWaitTime);
-                mySparky.StrafeRight(CyDogsChassis.OneTileMM-40, .5, mySparky.StandardAutonWaitTime);
+                //Already facing the correct way
+                //We're 'BackUpDistanceFromSpike' closer to scoreboard
+                mySparky.RotateLeft(3,.5,mySparky.StandardAutonWaitTime);
+                mySparky.StrafeLeft(40,.5,mySparky.StandardAutonWaitTime);
                 mySparky.raiseArmToScore(CyDogsSparky.ArmRaiseBeforeElbowMovement);
-                // has long wait time to handle arm movement before it
-                mySparky.RotateRight(188, .5, 2000);
+                mySparky.MoveStraight(675, .5, 2000);
+                // I took 200 off the above to be far enough away to read april tags
+            }
 
-             }
 
-
-            mySparky.AdjustToAprilTag(mySpikeLocation,"RedRight");
+            mySparky.AdjustToAprilTag(mySpikeLocation,"BlueLeft");
             mySparky.scoreFromDrivingPositionAndReturn();
             mySparky.MoveStraight(-50,.5,300);
             mySparky.AutonParkInCorrectSpot(mySpikeLocation, parkingSpot);
             mySparky.returnArmFromScoring();
             mySparky.LowerArmAtAutonEnd();
             //mySparky.MoveStraight(100,.5,300);
-
         }
     }
 
 
-
-
 }
-
