+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/centerstage/AutonBlueRight.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.centerstage;
 
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
@@ -6,10 +6,10 @@ import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.util.ElapsedTime;
 
 @Autonomous
-public class AutonRedLeft extends LinearOpMode {
+public class AutonBlueRight extends LinearOpMode {
     SpikeCam.location mySpikeLocation;
 
-// This is a LONG side Auton
+    // This is a LONG side Auton
     @Override
     public void runOpMode() {
         ElapsedTime runtime = new ElapsedTime();
@@ -18,18 +18,18 @@ public class AutonRedLeft extends LinearOpMode {
 
         // Set defaults for initialization options
         CyDogsChassis.Direction parkingSpot = CyDogsChassis.Direction.LEFT;
-        CyDogsChassis.Direction drivePath = CyDogsChassis.Direction.LEFT;
+        CyDogsChassis.Direction drivePath = CyDogsChassis.Direction.RIGHT;
 
         // Create the instance of sparky, initialize the SpikeCam, devices, and positions
-        CyDogsSparky mySparky = new CyDogsSparky(this, CyDogsChassis.Alliance.RED, 430);
+        CyDogsSparky mySparky = new CyDogsSparky(this, CyDogsChassis.Alliance.BLUE, 440);
         mySparky.initializeSpikeCam();
         mySparky.initializeDevices();
-
+     //   mySparky.initializePositions();
         mySparky.initializeAprilTags();
 
         // Ask the initialization questions
         parkingSpot = mySparky.askParkingSpot();
-     //   drivePath = mySparky.askDrivePath();
+       // drivePath = mySparky.askDrivePath();
 
         // Wait for the start button to be pressed on the driver station
         waitForStart();
@@ -47,25 +47,24 @@ public class AutonRedLeft extends LinearOpMode {
 
             if(mySpikeLocation==SpikeCam.location.LEFT)
             {
-                mySparky.StrafeLeft(100,0.5, mySparky.StandardAutonWaitTime);
+                mySparky.StrafeRight(80,0.5, mySparky.StandardAutonWaitTime);
             }
             else if(mySpikeLocation==SpikeCam.location.MIDDLE) {
-                mySparky.StrafeLeft(100,0.5, mySparky.StandardAutonWaitTime);
+                mySparky.StrafeRight(100,0.5, mySparky.StandardAutonWaitTime);
             }
             else {  //RIGHT
-                mySparky.StrafeLeft(90,0.5, mySparky.StandardAutonWaitTime);
+                mySparky.StrafeRight(85,0.5, mySparky.StandardAutonWaitTime);
             }
 
             mySparky.MoveStraight(-445, .5, mySparky.StandardAutonWaitTime);
-
 
             // Place purple pixel and back away from it
             mySparky.AutonPlacePurplePixel(mySpikeLocation);
             int backAwayFromPurple = 20;
             if(mySpikeLocation== SpikeCam.location.MIDDLE){
-                backAwayFromPurple=85;
+                backAwayFromPurple=75;
             }
-            else if(mySpikeLocation==SpikeCam.location.RIGHT)
+            else if(mySpikeLocation==SpikeCam.location.LEFT)
             {
                 backAwayFromPurple=20+CyDogsSparky.BackUpDistanceFromSpike+50;
             }
@@ -74,47 +73,41 @@ public class AutonRedLeft extends LinearOpMode {
 
             switch (mySpikeLocation) {
                 case LEFT:
-                    extraVerticalMovement-=50;
-                //    mySparky.MoveStraight(mySparky.BackUpDistanceFromSpike,.5,mySparky.StandardAutonWaitTime);
-               //    mySparky.raiseArmToScore(400);
-                //    sleep(300);
-                    mySparky.StrafeLeft(CyDogsSparky.OneTileMM, .5, mySparky.StandardAutonWaitTime);
-                //    mySparky.RotateRight(2,.5,mySparky.StandardAutonWaitTime);
-               //     mySparky.raiseArmToScore(0);
+                case NOT_FOUND:
+                    extraVerticalMovement+=70;
+                    mySparky.RotateRight(190,.5,mySparky.StandardAutonWaitTime);
+
+                    mySparky.StrafeRight(CyDogsSparky.OneTileMM,.5,mySparky.StandardAutonWaitTime);
                     break;
                 case MIDDLE:
 
-                    mySparky.RotateLeft(91,.5,mySparky.StandardAutonWaitTime);
+                    mySparky.RotateRight(91,.5,mySparky.StandardAutonWaitTime);
 
-                    mySparky.MoveStraight(-CyDogsChassis.OneTileMM+180,.5,mySparky.StandardAutonWaitTime);
-                    mySparky.StrafeLeft(CyDogsSparky.OneTileMM+240,.5,500);
-                    //mySparky.MoveStraight(CyDogsChassis.OneTileMM,.5,mySparky.StandardAutonWaitTime);
+                    mySparky.MoveStraight(-CyDogsChassis.OneTileMM+220,.5,mySparky.StandardAutonWaitTime);
+                    mySparky.StrafeRight(CyDogsSparky.OneTileMM+240,.5,mySparky.StandardAutonWaitTime);
+                    mySparky.RotateLeft(2,.5,mySparky.StandardAutonWaitTime);
                     extraVerticalMovement+=410;
 
                     break;
                 case RIGHT:
-                case NOT_FOUND:
-                    extraVerticalMovement+=70;
-                    mySparky.RotateRight(190,.5,mySparky.StandardAutonWaitTime);
-                    mySparky.StrafeLeft(CyDogsSparky.OneTileMM,.5,mySparky.StandardAutonWaitTime);
+                    extraVerticalMovement-=50;
+                    mySparky.raiseArmToScore(400);
+                    sleep(300);
+                    mySparky.StrafeRight(CyDogsSparky.OneTileMM+10, .5, mySparky.StandardAutonWaitTime);
+                    mySparky.raiseArmToScore(0);
+                    mySparky.RotateLeft(2,.5,mySparky.StandardAutonWaitTime);
 
                     break;
-
             }
             // We move not all the way so we don't crash into a parker, then after strafe move the rest
-
-            mySparky.MoveStraight(1900+extraVerticalMovement,.5,1000);
+            mySparky.MoveStraight(1900+extraVerticalMovement,.5,500);
             if(mySpikeLocation== SpikeCam.location.MIDDLE) {
-                mySparky.StrafeRight(75,.5,mySparky.StandardAutonWaitTime);
-            }
-            if(mySpikeLocation== SpikeCam.location.LEFT) {
-                mySparky.StrafeRight(75,.5,mySparky.StandardAutonWaitTime);
+                mySparky.StrafeLeft(100,.5,mySparky.StandardAutonWaitTime);
             }
             mySparky.raiseArmToScore(CyDogsSparky.ArmRaiseBeforeElbowMovement);
             mySparky.AutonCenterOnScoreboardBasedOnPath(drivePath);
 
-            mySparky.AdjustToAprilTag(mySpikeLocation,"RedLeft");
-            mySparky.MoveStraight(10, .5, mySparky.StandardAutonWaitTime);
+            mySparky.AdjustToAprilTag(mySpikeLocation,"BlueRight");
             mySparky.scoreFromDrivingPositionAndReturn();
             mySparky.MoveStraight(-50,.5,300);
             mySparky.AutonParkInCorrectSpot(mySpikeLocation, parkingSpot);
