+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/centerstage/AutonRedLeft.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.centerstage;
 
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
@@ -6,10 +6,10 @@ import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.util.ElapsedTime;
 
 @Autonomous
-public class AutonBlueRight extends LinearOpMode {
+public class AutonRedLeft extends LinearOpMode {
     SpikeCam.location mySpikeLocation;
 
-    // This is a LONG side Auton
+// This is a LONG side Auton
     @Override
     public void runOpMode() {
         ElapsedTime runtime = new ElapsedTime();
@@ -18,18 +18,18 @@ public class AutonBlueRight extends LinearOpMode {
 
         // Set defaults for initialization options
         CyDogsChassis.Direction parkingSpot = CyDogsChassis.Direction.LEFT;
-        CyDogsChassis.Direction drivePath = CyDogsChassis.Direction.RIGHT;
+        CyDogsChassis.Direction drivePath = CyDogsChassis.Direction.LEFT;
 
         // Create the instance of sparky, initialize the SpikeCam, devices, and positions
-        CyDogsSparky mySparky = new CyDogsSparky(this, CyDogsChassis.Alliance.BLUE, 440);
+        CyDogsSparky mySparky = new CyDogsSparky(this, CyDogsChassis.Alliance.RED, 430);
         mySparky.initializeSpikeCam();
         mySparky.initializeDevices();
-     //   mySparky.initializePositions();
+
         mySparky.initializeAprilTags();
 
         // Ask the initialization questions
         parkingSpot = mySparky.askParkingSpot();
-       // drivePath = mySparky.askDrivePath();
+     //   drivePath = mySparky.askDrivePath();
 
         // Wait for the start button to be pressed on the driver station
         waitForStart();
@@ -47,24 +47,25 @@ public class AutonBlueRight extends LinearOpMode {
 
             if(mySpikeLocation==SpikeCam.location.LEFT)
             {
-                mySparky.StrafeRight(80,0.5, mySparky.StandardAutonWaitTime);
+                mySparky.StrafeLeft(100,0.5, mySparky.StandardAutonWaitTime);
             }
             else if(mySpikeLocation==SpikeCam.location.MIDDLE) {
-                mySparky.StrafeRight(100,0.5, mySparky.StandardAutonWaitTime);
+                mySparky.StrafeLeft(100,0.5, mySparky.StandardAutonWaitTime);
             }
             else {  //RIGHT
-                mySparky.StrafeRight(85,0.5, mySparky.StandardAutonWaitTime);
+                mySparky.StrafeLeft(90,0.5, mySparky.StandardAutonWaitTime);
             }
 
             mySparky.MoveStraight(-445, .5, mySparky.StandardAutonWaitTime);
+
 
             // Place purple pixel and back away from it
             mySparky.AutonPlacePurplePixel(mySpikeLocation);
             int backAwayFromPurple = 20;
             if(mySpikeLocation== SpikeCam.location.MIDDLE){
-                backAwayFromPurple=75;
+                backAwayFromPurple=85;
             }
-            else if(mySpikeLocation==SpikeCam.location.LEFT)
+            else if(mySpikeLocation==SpikeCam.location.RIGHT)
             {
                 backAwayFromPurple=20+CyDogsSparky.BackUpDistanceFromSpike+50;
             }
@@ -73,41 +74,47 @@ public class AutonBlueRight extends LinearOpMode {
 
             switch (mySpikeLocation) {
                 case LEFT:
-                case NOT_FOUND:
-                    extraVerticalMovement+=70;
-                    mySparky.RotateRight(190,.5,mySparky.StandardAutonWaitTime);
-
-                    mySparky.StrafeRight(CyDogsSparky.OneTileMM,.5,mySparky.StandardAutonWaitTime);
+                    extraVerticalMovement-=50;
+                //    mySparky.MoveStraight(mySparky.BackUpDistanceFromSpike,.5,mySparky.StandardAutonWaitTime);
+               //    mySparky.raiseArmToScore(400);
+                //    sleep(300);
+                    mySparky.StrafeLeft(CyDogsSparky.OneTileMM, .5, mySparky.StandardAutonWaitTime);
+                //    mySparky.RotateRight(2,.5,mySparky.StandardAutonWaitTime);
+               //     mySparky.raiseArmToScore(0);
                     break;
                 case MIDDLE:
 
-                    mySparky.RotateRight(91,.5,mySparky.StandardAutonWaitTime);
+                    mySparky.RotateLeft(91,.5,mySparky.StandardAutonWaitTime);
 
-                    mySparky.MoveStraight(-CyDogsChassis.OneTileMM+220,.5,mySparky.StandardAutonWaitTime);
-                    mySparky.StrafeRight(CyDogsSparky.OneTileMM+240,.5,mySparky.StandardAutonWaitTime);
-                    mySparky.RotateLeft(2,.5,mySparky.StandardAutonWaitTime);
+                    mySparky.MoveStraight(-CyDogsChassis.OneTileMM+180,.5,mySparky.StandardAutonWaitTime);
+                    mySparky.StrafeLeft(CyDogsSparky.OneTileMM+240,.5,500);
+                    //mySparky.MoveStraight(CyDogsChassis.OneTileMM,.5,mySparky.StandardAutonWaitTime);
                     extraVerticalMovement+=410;
 
                     break;
                 case RIGHT:
-                    extraVerticalMovement-=50;
-                    mySparky.raiseArmToScore(400);
-                    sleep(300);
-                    mySparky.StrafeRight(CyDogsSparky.OneTileMM+10, .5, mySparky.StandardAutonWaitTime);
-                    mySparky.raiseArmToScore(0);
-                    mySparky.RotateLeft(2,.5,mySparky.StandardAutonWaitTime);
+                case NOT_FOUND:
+                    extraVerticalMovement+=70;
+                    mySparky.RotateRight(190,.5,mySparky.StandardAutonWaitTime);
+                    mySparky.StrafeLeft(CyDogsSparky.OneTileMM,.5,mySparky.StandardAutonWaitTime);
 
                     break;
+
             }
             // We move not all the way so we don't crash into a parker, then after strafe move the rest
-            mySparky.MoveStraight(1900+extraVerticalMovement,.5,500);
+
+            mySparky.MoveStraight(1900+extraVerticalMovement,.5,1000);
             if(mySpikeLocation== SpikeCam.location.MIDDLE) {
-                mySparky.StrafeLeft(100,.5,mySparky.StandardAutonWaitTime);
+                mySparky.StrafeRight(75,.5,mySparky.StandardAutonWaitTime);
+            }
+            if(mySpikeLocation== SpikeCam.location.LEFT) {
+                mySparky.StrafeRight(75,.5,mySparky.StandardAutonWaitTime);
             }
             mySparky.raiseArmToScore(CyDogsSparky.ArmRaiseBeforeElbowMovement);
             mySparky.AutonCenterOnScoreboardBasedOnPath(drivePath);
 
-            mySparky.AdjustToAprilTag(mySpikeLocation,"BlueRight");
+            mySparky.AdjustToAprilTag(mySpikeLocation,"RedLeft");
+            mySparky.MoveStraight(10, .5, mySparky.StandardAutonWaitTime);
             mySparky.scoreFromDrivingPositionAndReturn();
             mySparky.MoveStraight(-50,.5,300);
             mySparky.AutonParkInCorrectSpot(mySpikeLocation, parkingSpot);
