+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/centerstage/AutonRedRightTest.java
@@ -1,19 +1,16 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.centerstage;
 
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.util.ElapsedTime;
 
-import org.firstinspires.ftc.vision.apriltag.AprilTagDetection;
 
 @Autonomous
-public class AutonBlueLeft extends LinearOpMode {
-// This is a SHORT side Auton
-SpikeCam.location mySpikeLocation;
-    CyDogsSparky mySparky;
+public class AutonRedRightTest extends LinearOpMode {
+    SpikeCam.location mySpikeLocation;
 
-    // This is a SHORT side Auton
+    // This is SHORT side Auton test
     @Override
     public void runOpMode() {
 
@@ -25,10 +22,10 @@ SpikeCam.location mySpikeLocation;
         CyDogsChassis.Direction parkingSpot = CyDogsChassis.Direction.LEFT;
 
         // Create the instance of sparky, initialize the SpikeCam, devices, and positions
-        mySparky = new CyDogsSparky(this, CyDogsChassis.Alliance.BLUE,350);
+        CyDogsSparky mySparky = new CyDogsSparky(this, CyDogsChassis.Alliance.RED, 330);
         mySparky.initializeSpikeCam();
         mySparky.initializeDevices();
-
+ //       mySparky.initializePositions();
         mySparky.initializeAprilTags();
 
         // Ask the initialization questions
@@ -45,58 +42,63 @@ SpikeCam.location mySpikeLocation;
 
             // Get to standard position before placing purple pixel
             mySparky.MoveStraight(-300, .5, mySparky.StandardAutonWaitTime);
-            mySparky.StrafeLeft(75,0.5, mySparky.StandardAutonWaitTime);
+            mySparky.StrafeRight(90,0.5, mySparky.StandardAutonWaitTime);
             mySparky.MoveStraight(-445, .5, mySparky.StandardAutonWaitTime);
 
             // Place purple pixel and back away from it
             mySparky.AutonPlacePurplePixel(mySpikeLocation);
-            if(mySpikeLocation== SpikeCam.location.LEFT) {
-                mySparky.MoveStraight(65, .5, mySparky.StandardAutonWaitTime);
+            if(mySpikeLocation== SpikeCam.location.MIDDLE) {
                 mySparky.raiseArmToScore(CyDogsSparky.ArmRaiseBeforeElbowMovement);
-                sleep(400);
+                mySparky.MoveStraight(65, .5, mySparky.StandardAutonWaitTime);
             }
             else {
-                mySparky.MoveStraight(30, .5, mySparky.StandardAutonWaitTime);
+                mySparky.MoveStraight(20, .5, mySparky.StandardAutonWaitTime);
             }
-
 
             // First, let's get ourselves straight facing scoring area
             //   Then, adjust position.  Remember dropping purple pixel moved us back from spike 20mm
             if (mySpikeLocation == SpikeCam.location.LEFT) {
-
-                mySparky.StrafeRight(CyDogsChassis.OneTileMM, .5, mySparky.StandardAutonWaitTime);
-                mySparky.MoveStraight(-CyDogsChassis.OneTileMM-160, .5, mySparky.StandardAutonWaitTime);
-                mySparky.StrafeLeft(CyDogsChassis.OneTileMM-40, .5, mySparky.StandardAutonWaitTime);
-
-                mySparky.RotateLeft(188, .5, 2000);
-            } else if (mySpikeLocation == SpikeCam.location.MIDDLE) {
-                mySparky.RotateRight(92, .5, mySparky.StandardAutonWaitTime);
-                // We're 50mm further away from start position
-                mySparky.StrafeLeft(-50,.5,mySparky.StandardAutonWaitTime);
-                mySparky.raiseArmToScore(CyDogsSparky.ArmRaiseBeforeElbowMovement);
-                mySparky.MoveStraight(500, .5, 2000);
-            } else {  //RIGHT
                 //Already facing the correct way
                 //We're 'BackUpDistanceFromSpike' closer to scoreboard
-                mySparky.RotateLeft(3,.5,mySparky.StandardAutonWaitTime);
-                mySparky.StrafeLeft(40,.5,mySparky.StandardAutonWaitTime);
+                mySparky.RotateRight(3,.5,mySparky.StandardAutonWaitTime);
+                mySparky.StrafeRight(40,.5,mySparky.StandardAutonWaitTime);
                 mySparky.raiseArmToScore(CyDogsSparky.ArmRaiseBeforeElbowMovement);
+                // has long wait time to handle arm movement before it
                 mySparky.MoveStraight(675, .5, 2000);
-                // I took 200 off the above to be far enough away to read april tags
-            }
+      // I took 200 off the above to be far enough away to read april tags
+            } else if (mySpikeLocation == SpikeCam.location.MIDDLE) {
+                mySparky.RotateLeft(92, .5, mySparky.StandardAutonWaitTime);
+                // We're 50mm further away from start position
+                mySparky.StrafeRight(-50,.5,mySparky.StandardAutonWaitTime);
+               // mySparky.raiseArmToScore(CyDogsSparky.ArmRaiseBeforeElbowMovement);
+                // has long wait time to handle arm movement before it
+                mySparky.MoveStraight(500, .5, 2000);
+            } else {  //RIGHT
+                mySparky.StrafeLeft(CyDogsChassis.OneTileMM, .5, mySparky.StandardAutonWaitTime);
+                mySparky.MoveStraight(-CyDogsChassis.OneTileMM-160, .5, mySparky.StandardAutonWaitTime);
+                mySparky.StrafeRight(CyDogsChassis.OneTileMM-40, .5, mySparky.StandardAutonWaitTime);
+                mySparky.raiseArmToScore(CyDogsSparky.ArmRaiseBeforeElbowMovement);
+                // has long wait time to handle arm movement before it
+                mySparky.RotateRight(188, .5, 2000);
+
+             }
 
 
-            mySparky.AdjustToAprilTag(mySpikeLocation,"BlueLeft");
+            mySparky.AdjustToAprilTag(mySpikeLocation,"RedRight");
             mySparky.scoreFromDrivingPositionAndReturn();
             mySparky.MoveStraight(-50,.5,300);
             mySparky.AutonParkInCorrectSpot(mySpikeLocation, parkingSpot);
             mySparky.returnArmFromScoring();
             mySparky.LowerArmAtAutonEnd();
             //mySparky.MoveStraight(100,.5,300);
+
         }
     }
 
 
+
+
 }
+
