+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/centerstage/Learning_Auton_3.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.centerstage;
 
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
